bots/Tor_Spider/requirements.txt
beautifulsoup4==4.6.0
PySocks==1.6.7
stem==1.5.4
tldextract